+++ MoPubSDK/Internal/Utility/MPReachability.m
@@ -119,6 +119,8 @@ static void ReachabilityCallback(SCNetworkReachabilityRef target, SCNetworkReach
         {
             returnValue->_reachabilityRef = reachability;
             returnValue->_alwaysReturnLocalWiFiStatus = NO;
+        } else {
+            CFRelease(reachability);
         }
     }
     return returnValue;
@@ -138,6 +140,8 @@ static void ReachabilityCallback(SCNetworkReachabilityRef target, SCNetworkReach
         {
             returnValue->_reachabilityRef = reachability;
             returnValue->_alwaysReturnLocalWiFiStatus = NO;
+        } else {
+            CFRelease(reachability);
         }
     }
     return returnValue;