+++ Specs/MoPubSDKSpecs/Internal/Banners/MPBannerAdManagerSpec.mm
@@ -54,7 +54,7 @@ describe(@"MPBannerAdManager", ^{
             URL should contain(@"id=panther");
             URL should contain(@"q=liono");
             URL should contain(@"ll=30,20");
-            URL should contain(@"http://testing.ads.mopub.com");
+            URL should contain(@"https://testing.ads.mopub.com");
         });
     });