+++ Specs/MoPubSDKSpecs/Internal/Common/MPAdServerURLBuilderSpec.mm
@@ -36,7 +36,7 @@ describe(@"MPAdServerURLBuilder", ^{
                                                keywords:nil
                                                location:nil
                                                 testing:YES];
-            expected = [NSString stringWithFormat:@"http://testing.ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
+            expected = [NSString stringWithFormat:@"https://testing.ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
                         [MPIdentityProvider identifier],
                         MP_SDK_VERSION];
             URL.absoluteString should contain(expected);
@@ -45,7 +45,7 @@ describe(@"MPAdServerURLBuilder", ^{
                                                keywords:nil
                                                location:nil
                                                 testing:NO];
-            expected = [NSString stringWithFormat:@"http://ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
+            expected = [NSString stringWithFormat:@"https://ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
                         [MPIdentityProvider identifier],
                         MP_SDK_VERSION];
             URL.absoluteString should contain(expected);
@@ -390,21 +390,21 @@ describe(@"MPAdServerURLBuilder", ^{
         });
     });
 
-    context(@"when HTTPS is enabled", ^{
+    context(@"when HTTPS is disabled", ^{
         beforeEach(^{
-            [MPAPIEndpoints setUsesHTTPS:YES];
-        });
-
-        afterEach(^{
             [MPAPIEndpoints setUsesHTTPS:NO];
         });
 
-        it(@"should return HTTPS URLs", ^{
+        afterEach(^{
+            [MPAPIEndpoints setUsesHTTPS:YES];
+        });
+
+        it(@"should return HTTP URLs", ^{
             URL = [MPAdServerURLBuilder URLWithAdUnitID:@"guy"
                                                keywords:nil
                                                location:nil
                                                 testing:NO];
-            expected = [NSString stringWithFormat:@"https://ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
+            expected = [NSString stringWithFormat:@"http://ads.mopub.com/m/ad?v=8&udid=%@&id=guy&nv=%@",
                         [MPIdentityProvider identifier],
                         MP_SDK_VERSION];
             URL.absoluteString should contain(expected);