+++ Specs/MoPubSDKSpecs/Internal/Native/MPStaticNativeAdImpressionTimerSpec.mm
@@ -1,13 +1,20 @@
-#import "MPStaticNativeAdImpressionTimer.h"
+#import "MPStaticNativeAdImpressionTimer+Specs.h"
+#import "MPTimer.h"
 
 using namespace Cedar::Matchers;
 using namespace Cedar::Doubles;
 
 SPEC_BEGIN(MPStaticNativeAdImpressionTimerSpec)
 
-xdescribe(@"MPStaticNativeAdImpressionTimerSpec", ^{
-    beforeEach(^{
+describe(@"MPStaticNativeAdImpressionTimerSpec", ^{
+    __block MPStaticNativeAdImpressionTimer *impressionTimer;
 
+    beforeEach(^{
+        impressionTimer = [[MPStaticNativeAdImpressionTimer alloc] initWithRequiredSecondsForImpression:1 requiredViewVisibilityPercentage:.5];
+    });
+
+    it(@"should default first visibility time stamp to -1", ^{
+        impressionTimer.firstVisibilityTimestamp should equal(-1);
     });
 });