+++ Specs/MoPubSDKSpecs/MPAdConversionTrackerSpec.mm
@@ -32,7 +32,7 @@ describe(@"MPAdConversionTracker", ^{
             [request valueForHTTPHeaderField:@"User-Agent"] should equal(@"FAKE_TEST_USER_AGENT_STRING");
 
             NSString *URL = request.URL.absoluteString;
-            URL should contain(@"http://ads.mopub.com/m/open?v=8");
+            URL should contain(@"https://ads.mopub.com/m/open?v=8");
             URL should contain(@"&id=128405");
             URL should contain(@"&av=1.0");
 
@@ -83,22 +83,22 @@ describe(@"MPAdConversionTracker", ^{
         });
     });
 
-    context(@"when HTTPS is enabled", ^{
+    context(@"when HTTPS is disabled", ^{
         beforeEach(^{
-            [MPAPIEndpoints setUsesHTTPS:YES];
-        });
-
-        afterEach(^{
             [MPAPIEndpoints setUsesHTTPS:NO];
         });
 
-        it(@"should make its API call over HTTPS", ^{
+        afterEach(^{
+            [MPAPIEndpoints setUsesHTTPS:YES];
+        });
+
+        it(@"should make its API call over HTTP", ^{
             [tracker reportApplicationOpenForApplicationID:applicationID];
             NSURLConnection *connection = [[NSURLConnection connections] lastObject];
 
             NSURLRequest *request = [connection request];
             NSString *URL = request.URL.absoluteString;
-            URL should contain(@"https://ads.mopub.com/m/open?v=8");
+            URL should contain(@"http://ads.mopub.com/m/open?v=8");
         });
     });
 });