+++ Specs/MoPubSDKSpecs/MPAdViewSpec.mm
@@ -41,7 +41,7 @@ describe(@"MPAdView", ^{
             requestedPath should contain(@"&q=hi=4");
             requestedPath should contain(@"&ll=20,20");
             requestedPath should contain(@"&lla=100");
-            requestedPath should contain(@"http://testing.ads.mopub.com");
+            requestedPath should contain(@"https://testing.ads.mopub.com");
         });
     });