+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 4.4 (February 17th, 2016)
+
+- **Enabled SSL ad requests by default**
+- **Bug fixes**
+	- Fixed native video impression tracking
+	- Made closeable hot spot consistent across all full-screen creative types
+	- Added boundry checks to address https://github.com/mopub/mopub-ios-sdk/pull/138
+
+
 ## Version 4.3 (December 15th, 2015)
 
 - **Minor SDK improvements**.